bluetooth: Cancelled CG2900 UART works
This patch fixes an issue in the CG2900 UART driver where a chip enable
sometimes failed due to that the sleep_state had been set to an incorrect
state. This was due to a work that was executed after chip had been
shut down.

ST-Ericsson Linux next: Not tested, ER282957
ST-Ericsson ID: 319624
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I9043db9fc480ef20f677674e646654984c3199e6
Signed-off-by: Par-Gunnar Hjalmdahl <par-gunnar.p.hjalmdahl@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/12484
Reviewed-by: QATOOLS