Firmware: remove CG2900 firmware from the kernel.
This patch removes the firmware files for CG2900 chip from the
kernel/firmware folder and modifies the firmware/Makefile so that they
will not be built-in the kernel binary. They will be placed in a new
git kernel-firmware and loaded in runtime using the firmware_class.
It also modifies the way CG2900 driver determines which firmware
to load - instead of parse a text files the file name is created
using information obtained from the connected chip.

Depends-On: I53c8d6928b7d0c1fe08732c61c05c015bee92bf

ST-Ericsson ID: CR 278033
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I2cbd4efa3f03353c457997248aaba9a25a320151
Signed-off-by: Dariusz Szymczak <dariusz.xd.szymczak@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/13190
Reviewed-by: QATOOLS
Reviewed-by: Par-Gunnar HJALMDAHL <par-gunnar.p.hjalmdahl@stericsson.com>